asyn dualport ram
Inputs and Outputs:

wr_clk and rd_clk: These are the write and read clocks, respectively. Data is written to the RAM on the rising edge of the write clock (wr_clk) and read from the RAM on the rising edge of the read clock (rd_clk).
clr: This is a clear signal that resets the RAM contents to zero when asserted (clr = 1'b1).
we: This is the write enable signal. When asserted (we = 1'b1), it allows data to be written to the RAM.
re: This is the read enable signal. When asserted (re = 1'b1), it allows data to be read from the RAM.
wr_addr and rd_addr: These are the write and read addresses, respectively. They specify the location in the RAM where data is written to or read from.
data_in: This is the 16-bit data input that is written into the RAM when we is active.
data_out: This is the 16-bit data output from the RAM when re is active.
RAM Array:

reg [15:0] ram[7:0];: This declares an array ram with 8 elements, each being a 16-bit wide register. These registers represent the memory locations in the RAM.
Clearing the RAM:

When the clr signal is asserted (clr = 1'b1), the RAM is cleared. The always block with @(posedge wr_clk or posedge clr) is triggered either on a rising edge of wr_clk or when clr is asserted.
Inside this block, a for loop is used to iterate through all 8 elements of the ram array and set them to zero (ram[i] <= 0;).
Writing to the RAM:

Data can be written to the RAM when we is asserted (we = 1'b1).
The always block with @(posedge wr_clk or posedge clr) is triggered on a rising edge of wr_clk.
If clr is not asserted (i.e., the RAM is not being cleared), and we is active, the data specified by data_in is written into the location specified by wr_addr in the ram array (ram[wr_addr] <= data_in;).
Reading from the RAM:

Data can be read from the RAM when re is asserted (re = 1'b1).
The always block with @(posedge rd_clk or posedge clr) is triggered on a rising edge of rd_clk.
If clr is not asserted (i.e., the RAM is not being cleared), and re is active, the data from the location specified by rd_addr in the ram array is assigned to data_out (data_out = ram[rd_addr];).